back-end/pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Calender API tool"
authors = ["ananya-app"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.11.3"
fastapi = "~0.68.0"
SQLAlchemy = "^1.4.22"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.9.1"}
google-api-python-client = "^2.0.0"
google_auth_oauthlib = "^1.2.0"
starlette="^0.14.2"
itsdangerous="^2.1.2"
openai="^0.27.0"